requirements.txt
streamlit
supabase
st-supabase-connection
pandas
xlsxwriter
python-dotenv